% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aidR-package.R
\docType{package}
\name{aidR-package}
\alias{aidR}
\alias{aidR-package}
\title{aidR: Reading and Standardizing Automated Insulin Delivery Data}
\description{
Parses the native exports of several automated insulin delivery (AID) platforms into a common, platform-independent format for continuous glucose monitoring (CGM), basal, bolus, carbohydrate, and self-monitored blood glucose (SMBG) data. Additional functions check completeness, produce visualizations, and convert the data to other formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/madleina/aidR}
  \item Report bugs at \url{https://github.com/madleina/aidR/issues}
}

}
\author{
\strong{Maintainer}: Madleina Caduff \email{madleina.caduff@dcberne.com} (\href{https://orcid.org/0000-0001-8941-4274}{ORCID})

Authors:
\itemize{
  \item Madleina Caduff \email{madleina.caduff@dcberne.com} (\href{https://orcid.org/0000-0001-8941-4274}{ORCID})
  \item Aritz Lizoain \email{aritz.lizoain@dcberne.com} (\href{https://orcid.org/0009-0004-5516-2573}{ORCID})
}

Other contributors:
\itemize{
  \item Diabetes Center Berne (DCB) [copyright holder, funder]
}

}
\keyword{internal}
