% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_checks.R
\name{check_completeness_range}
\alias{check_completeness_range}
\title{Check if data for different types are available for a range of dates}
\usage{
check_completeness_range(
  id,
  data_id,
  first_date,
  last_date,
  necessary_types = c("cgm", "basal", "bolus", "carbs")
)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data_id}{A list with all data found for one individual.}

\item{first_date}{A character string representing the first date for which data should be available.}

\item{last_date}{A character string representing the last date (included) for which data should be available.}

\item{necessary_types}{A character vector with type names for which data should be available.}
}
\value{
No return value, called for side effects. Issues a warning for every
data type with missing days.
}
\description{
Check if data for different types are available for a range of dates
}
