% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{clean.carelink}
\alias{clean.carelink}
\title{Format and clean the CareLink data to keep relevant columns only}
\usage{
\method{clean}{carelink}(data, id, ...)
}
\arguments{
\item{data}{An instance of class \code{carelink}, wrapping a named list with data of different types.}

\item{id}{Character or numeric participant identifier.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with the cleaned data, with entries \code{cgm},
\code{bolus}, \code{carbs} (extracted from the bolus wizard
entries) and \code{SMBG}.
}
\description{
Format and clean the CareLink data to keep relevant columns only
}
\keyword{internal}
