% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{clean.mylife}
\alias{clean.mylife}
\title{Format and clean the Mylife data to keep relevant columns only.}
\usage{
\method{clean}{mylife}(data, id, ...)
}
\arguments{
\item{data}{An instance of class \code{mylife}, wrapping a named list with data of different types.}

\item{id}{Character or numeric participant identifier.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with names \code{bolus} and \code{carbs}, and other names for data of other types.
}
\description{
Format and clean the Mylife data to keep relevant columns only.
}
