% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAidData.R
\name{clean_aidData}
\alias{clean_aidData}
\title{Clean China}
\usage{
clean_aidData(df, dataset = "Core", level = "full")
}
\arguments{
\item{df}{a Data Frame containing the Raw  Dataset}

\item{dataset}{Character String denoting which dataset is to be cleaned (takes "Core" or "China")}

\item{level}{Level (c("full","donor-year","donor-year-recipient","donor-year-recipient-type"))}
}
\value{
A data frame containing \code{n} rows of blabla.
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
x <- c(1,2,3)
mean(x)
}
