% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAidData.R
\name{clean_china}
\alias{clean_china}
\title{Clean China}
\usage{
clean_china(df, level = "full")
}
\arguments{
\item{df}{a Data Frame containing the Raw China Dataset}

\item{par2}{Level (c("full","donor-year","donor-year-recipient","donor-year-recipient-type"))}
}
\value{
A data frame containing \code{n} rows of blabla.
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
x <- c(1,2,3)
mean(x)
}
