% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_checks.R
\name{.check_completeness_id.aidR}
\alias{.check_completeness_id.aidR}
\title{Check if data for different types are available for at least one day, for one individual}
\usage{
.check_completeness_id.aidR(id, data_id, necessary_types)
}
\arguments{
\item{id}{Character participant identifier, or \code{NA} if it is unknown.}

\item{data_id}{A list with all data found for one individual.}

\item{necessary_types}{A character vector with type names for which data should be available.}
}
\value{
No return value, called for side effects.
}
\description{
Check if data for different types are available for at least one day, for one individual
}
\keyword{internal}
