% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.check_if_missing_columns_with_data.aidR}
\alias{.check_if_missing_columns_with_data.aidR}
\title{Check that no relevant columns were missed when splitting a data frame}
\usage{
.check_if_missing_columns_with_data.aidR(df, subset)
}
\arguments{
\item{df}{The full data frame.}

\item{subset}{The subsetted data frame.}
}
\value{
Called for its side effects. Returns \code{NULL} invisibly.
}
\description{
Throws an error if there are columns with non-NA entries in \code{df} that are
not present in \code{subset}.
}
\keyword{internal}
