% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.clean_carelink_carbs.aidR}
\alias{.clean_carelink_carbs.aidR}
\title{Format and clean carbohydrate data from a CareLink export}
\usage{
.clean_carelink_carbs.aidR(id, bwz)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bwz}{A data frame containing the bolus wizard entries from a CareLink
file, which hold the carbohydrates entered into the bolus calculator.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data.
}
\description{
Format and clean carbohydrate data from a CareLink export
}
\keyword{internal}
