% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{.clean_mylife_SMBG.aidR}
\alias{.clean_mylife_SMBG.aidR}
\title{Format and clean SMBG data from a mylife export}
\usage{
.clean_mylife_SMBG.aidR(id, SMBG)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{SMBG}{A data frame containing the SMBG data from a mylife file.}
}
\value{
A data frame containing the formatted and cleaned SMBG data.
\code{NULL} if there is no SMBG data.
}
\description{
Format and clean SMBG data from a mylife export
}
\keyword{internal}
