% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_checks.R
\name{.conflicting_dates.aidR}
\alias{.conflicting_dates.aidR}
\title{Find dates that appear more than once with differing values}
\usage{
.conflicting_dates.aidR(dates, values)
}
\arguments{
\item{dates}{A vector of dates.}

\item{values}{A numeric vector of the same length as \code{dates}.}
}
\value{
A vector with the dates that carry more than one distinct value.
}
\description{
Find dates that appear more than once with differing values
}
\keyword{internal}
