% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.convert_BG_units_carelink.aidR}
\alias{.convert_BG_units_carelink.aidR}
\title{Convert blood glucose units to mg/dL in a CareLink file}
\usage{
.convert_BG_units_carelink.aidR(df)
}
\arguments{
\item{df}{A data frame obtained from reading a CareLink file.}
}
\value{
A data frame, with all mmol/L columns converted to mg/dL and renamed
accordingly.
}
\description{
Convert blood glucose units to mg/dL in a CareLink file
}
\keyword{internal}
