% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.convert_excel_serial.aidR}
\alias{.convert_excel_serial.aidR}
\title{Convert timestamps from a Excel file formatted as serial numbers.}
\usage{
.convert_excel_serial.aidR(serial)
}
\arguments{
\item{serial}{The serial value.}
}
\value{
A proper timestamp in POSIXct.
}
\description{
Convert timestamps from a Excel file formatted as serial numbers.
}
\keyword{internal}
