% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.convert_timestamp_carelink.aidR}
\alias{.convert_timestamp_carelink.aidR}
\title{Convert timestamps of a CareLink file to POSIXct timestamps}
\usage{
.convert_timestamp_carelink.aidR(timestamp, date_val = NULL, time_val = NULL)
}
\arguments{
\item{timestamp}{The timestamp from the file.}

\item{date_val}{The fractional value for the date, used if timestamp can not be parsed directly.}

\item{time_val}{The fractional value for the time of day, used if timestamp can not be parsed directly.}
}
\value{
A vector of timestamps in POSIXct.
}
\description{
Convert timestamps of a CareLink file to POSIXct timestamps
}
\keyword{internal}
