% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.extract_estimated_absorption_duration_tidepool.aidR}
\alias{.extract_estimated_absorption_duration_tidepool.aidR}
\title{Extract the estimated absorption duration from the Food sheet of a Tidepool export}
\usage{
.extract_estimated_absorption_duration_tidepool.aidR(row, data)
}
\arguments{
\item{row}{The current row index to parse.}

\item{data}{The full data frame with the food data from a Tidepool file.}
}
\value{
A numeric value with the estimated absorption duration in seconds,
\code{NA} if not found.
}
\description{
The duration is either part of the JSON string in the \code{Nutrition} column,
or part of the JSON string in the \code{Payload} column.
}
\keyword{internal}
