% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.extract_net_carbs_tidepool.aidR}
\alias{.extract_net_carbs_tidepool.aidR}
\title{Extract net carbohydrate amount from json string given in "Nutrition" column of Food sheet}
\usage{
.extract_net_carbs_tidepool.aidR(json_str)
}
\arguments{
\item{json_str}{A character string in json format, containing the net carbohydrate amount}
}
\value{
A numeric value with the net carbohydrate amount
}
\description{
Extract net carbohydrate amount from json string given in "Nutrition" column of Food sheet
}
\keyword{internal}
