% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_filtering.R
\name{.filter_cgm.aidR}
\alias{.filter_cgm.aidR}
\title{Filter CGM data.}
\usage{
.filter_cgm.aidR(
  cgm,
  truncate_min_cgm = 40,
  truncate_max_cgm = 400,
  min_minutes_difference_cgm = 5
)
}
\arguments{
\item{cgm}{A data frame containing the CGM data for one individual.}

\item{truncate_min_cgm}{CGM values smaller than this threshold will be truncated. Default 40 mg/dL, if NULL, no filtering is performed.}

\item{truncate_max_cgm}{CGM values larger than this threshold will be truncated. Default 400 mg/dL, if NULL, no filtering is performed.}

\item{min_minutes_difference_cgm}{CGM values are downsampled to this time interval (in minutes). Default 5 minutes, if NULL, no downsampling is performed.}
}
\value{
A list with all data found for one individual, sorted in time
}
\description{
Filter CGM data.
}
\keyword{internal}
