% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_basal_glooko.aidR}
\alias{.format_basal_glooko.aidR}
\title{Format and clean basal data from Glooko.}
\usage{
.format_basal_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the basal data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned basal data.
}
\description{
Format and clean basal data from Glooko.
}
\keyword{internal}
