% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_bolus_calculator_tidepool.aidR}
\alias{.format_bolus_calculator_tidepool.aidR}
\title{Format and clean carbohydrate data from the Bolus Calculator sheet of a Tidepool export}
\usage{
.format_bolus_calculator_tidepool.aidR(id, bolus_calculator)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bolus_calculator}{A data frame containing the bolus calculator data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data.
\code{NULL} if no carbohydrates were entered.
}
\description{
Format and clean carbohydrate data from the Bolus Calculator sheet of a Tidepool export
}
\keyword{internal}
