% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_bolus_tidepool.aidR}
\alias{.format_bolus_tidepool.aidR}
\title{Format and clean bolus data from a Tidepool export}
\usage{
.format_bolus_tidepool.aidR(id, bolus)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bolus}{A data frame containing the bolus data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus data from a Tidepool export
}
\keyword{internal}
