% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_bolus_yourloops.aidR}
\alias{.format_bolus_yourloops.aidR}
\title{Format and clean bolus data from Yourloops}
\usage{
.format_bolus_yourloops.aidR(bolus)
}
\arguments{
\item{bolus}{A data frame containing the bolus data from a Yourloops file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus data from Yourloops
}
\keyword{internal}
