% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_carbs_glooko.aidR}
\alias{.format_carbs_glooko.aidR}
\title{Format and clean carbohydrate data from Glooko.}
\usage{
.format_carbs_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the carbohydrate data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data.
}
\description{
Format and clean carbohydrate data from Glooko.
}
\keyword{internal}
