% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.format_carbs_tandem_source.aidR}
\alias{.format_carbs_tandem_source.aidR}
\title{Format and clean carbs data from Tandem Source}
\usage{
.format_carbs_tandem_source.aidR(carbs)
}
\arguments{
\item{carbs}{A data frame containing the carbs data from a Tandem Source file.}
}
\value{
A data frame containing the formatted and cleaned carbs data.
}
\description{
Format and clean carbs data from Tandem Source
}
\keyword{internal}
