% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_carbs_yourloops.aidR}
\alias{.format_carbs_yourloops.aidR}
\title{Format and clean carbs data from Yourloops}
\usage{
.format_carbs_yourloops.aidR(id, carbs)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{carbs}{A data frame containing the carbs data from a Yourloops file.}
}
\value{
A data frame containing the formatted and cleaned carbs data.
}
\description{
Format and clean carbs data from Yourloops
}
\keyword{internal}
