% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01h_dexcom_clarity.R
\name{.format_cgm_dexcom_clarity.aidR}
\alias{.format_cgm_dexcom_clarity.aidR}
\title{Format and clean CGM data from a Dexcom Clarity export}
\usage{
.format_cgm_dexcom_clarity.aidR(id, cgm)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{cgm}{A data frame containing the CGM data from a Dexcom Clarity file.}
}
\value{
A data frame containing the formatted and cleaned CGM data.
}
\description{
Format and clean CGM data from a Dexcom Clarity export
}
\keyword{internal}
