% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.format_cgm_tandem_source.aidR}
\alias{.format_cgm_tandem_source.aidR}
\title{Format and clean CGM data from Tandem Source}
\usage{
.format_cgm_tandem_source.aidR(cgm)
}
\arguments{
\item{cgm}{A data frame containing the CGM data from a Tandem Source file.}
}
\value{
A data frame containing the formatted and cleaned CGM data.
}
\description{
Format and clean CGM data from Tandem Source
}
\keyword{internal}
