% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_cgm_tidepool.aidR}
\alias{.format_cgm_tidepool.aidR}
\title{Format and clean CGM data from a Tidepool export}
\usage{
.format_cgm_tidepool.aidR(id, cgm)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{cgm}{A data frame containing the CGM data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned CGM data.
}
\description{
Format and clean CGM data from a Tidepool export
}
\keyword{internal}
