% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_checks.R
\name{.format_date_ranges.aidR}
\alias{.format_date_ranges.aidR}
\title{Format date ranges}
\usage{
.format_date_ranges.aidR(dates)
}
\arguments{
\item{dates}{A vector with dates.}
}
\value{
A string with all consecutive dates collapsed into ranges.
}
\description{
Format date ranges
}
\keyword{internal}
