% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_conversions.R
\name{.format_diax_time.aidR}
\alias{.format_diax_time.aidR}
\title{Format timestamps according to the DIAX \code{"Y-m-d H:M:S Z"} convention.}
\usage{
.format_diax_time.aidR(timestamp, tz_offset)
}
\arguments{
\item{timestamp}{A vector of timestamps.}

\item{tz_offset}{Character string with the UTC offset of the timestamps, in the
form \code{"+HH:MM"} or \code{"-HH:MM"}.}
}
\value{
A character vector of formatted timestamps.
}
\description{
Format timestamps according to the DIAX \code{"Y-m-d H:M:S Z"} convention.
}
\keyword{internal}
