% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_exercise_glooko.aidR}
\alias{.format_exercise_glooko.aidR}
\title{Format and clean exercise data from Glooko.}
\usage{
.format_exercise_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the exercise data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned exercise data.
}
\description{
Format and clean exercise data from Glooko.
}
\keyword{internal}
