% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_manual_insulin_glooko.aidR}
\alias{.format_manual_insulin_glooko.aidR}
\title{Format and clean manual insulin data from Glooko.}
\usage{
.format_manual_insulin_glooko.aidR(data)
}
\arguments{
\item{data}{A data frame containing the manual insulin data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned manual insulin data.
}
\description{
Format and clean manual insulin data from Glooko.
}
\keyword{internal}
