% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_rescuecarbs_yourloops.aidR}
\alias{.format_rescuecarbs_yourloops.aidR}
\title{Format and clean rescue carbohydrate data from a YourLoops export}
\usage{
.format_rescuecarbs_yourloops.aidR(id, carbs)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{carbs}{A data frame containing the rescue carbohydrate data from a YourLoops file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data,
flagged as hypoglycemia treatment.
}
\description{
Format and clean rescue carbohydrate data from a YourLoops export
}
\keyword{internal}
