% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_timestamp_yourloops.aidR}
\alias{.format_timestamp_yourloops.aidR}
\title{Format timestamps to get local (wall-clock) time of a YourLoops export}
\usage{
.format_timestamp_yourloops.aidR(file)
}
\arguments{
\item{file}{A data frame obtained from reading a YourLoops file.}
}
\value{
A data frame where the timestamps have been converted to local time.
}
\description{
Timestamps are given in Zulu (UTC) time; the timezone offset is added to
obtain local time. Data frames without these two columns are left unchanged.
}
\keyword{internal}
