% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_bolus.aidR}
\alias{.get_carelink_bolus.aidR}
\title{Parse all entries corresponding to bolus deliveries in a Carelink export.}
\usage{
.get_carelink_bolus.aidR(df, pump_name)
}
\arguments{
\item{df}{A data frame obtained from reading a Carelink file.}

\item{pump_name}{A string denoting the name of the pump, NA if unknown.}
}
\value{
A data frame containing the target entries
}
\description{
Parse all entries corresponding to bolus deliveries in a Carelink export.
}
\keyword{internal}
