% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_carb_variant.aidR}
\alias{.get_carelink_carb_variant.aidR}
\title{Find the carbohydrate columns present in a CareLink export}
\usage{
.get_carelink_carb_variant.aidR(df)
}
\arguments{
\item{df}{A data frame containing the bolus wizard entries from a CareLink file.}
}
\value{
A list with the column names of the carbohydrate input (\code{input})
and the carbohydrate ratio (\code{ratio}), as well as the corresponding
unit (\code{unit}).
}
\description{
Find the carbohydrate columns present in a CareLink export
}
\keyword{internal}
