% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_scheduled_basal_rates.aidR}
\alias{.get_carelink_scheduled_basal_rates.aidR}
\title{Parse all entries corresponding to the scheduled basal rates in a CareLink export}
\usage{
.get_carelink_scheduled_basal_rates.aidR(df, pump_name)
}
\arguments{
\item{df}{A data frame obtained from reading a CareLink file.}

\item{pump_name}{A character string with the name of the pump, \code{NA} if unknown.}
}
\value{
A data frame with the corresponding entries.
}
\description{
Parse all entries corresponding to the scheduled basal rates in a CareLink export
}
\keyword{internal}
