% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01h_dexcom_clarity.R
\name{.get_sep_dexcom_clarity.aidR}
\alias{.get_sep_dexcom_clarity.aidR}
\title{Find the separator used in a Dexcom Clarity export}
\usage{
.get_sep_dexcom_clarity.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A character string with the separator (\code{";"} or \code{","}), or
\code{NULL} if the file does not come from a Dexcom Clarity export.
}
\description{
Find the separator used in a Dexcom Clarity export
}
\keyword{internal}
