% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.get_tandem_source_file_type.aidR}
\alias{.get_tandem_source_file_type.aidR}
\title{Get the data type of a file from a Tandem Source export}
\usage{
.get_tandem_source_file_type.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A character string with the data type, or \code{NULL} if there was no
match. Throws an error if multiple data types matched.
}
\description{
Get the data type of a file from a Tandem Source export
}
\keyword{internal}
