% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.get_yourloops_file_lookup.aidR}
\alias{.get_yourloops_file_lookup.aidR}
\title{A lookup with YourLoops file characteristics}
\usage{
.get_yourloops_file_lookup.aidR()
}
\value{
A list with data types and their expected headers.
}
\description{
A lookup with YourLoops file characteristics
}
\keyword{internal}
