% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01h_dexcom_clarity.R
\name{.is_dexcom_clarity_format.aidR}
\alias{.is_dexcom_clarity_format.aidR}
\title{Check if a file comes from a Dexcom Clarity export}
\usage{
.is_dexcom_clarity_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file comes from a Dexcom Clarity export.
}
\description{
Check if a file comes from a Dexcom Clarity export
}
\keyword{internal}
