% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.is_glooko_format.aidR}
\alias{.is_glooko_format.aidR}
\title{Check if a file comes from a Glooko export}
\usage{
.is_glooko_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file comes from a Glooko export.
}
\description{
Check if a file comes from a Glooko export
}
\keyword{internal}
