% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{.is_mylife_format.aidR}
\alias{.is_mylife_format.aidR}
\title{Check if a file comes from a mylife export}
\usage{
.is_mylife_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file comes from a mylife export.
}
\description{
Check if a file comes from a mylife export
}
\keyword{internal}
