% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.is_tidepool_format.aidR}
\alias{.is_tidepool_format.aidR}
\title{Check if a file corresponds to a file from a Tidepool export.}
\usage{
.is_tidepool_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file is a file from a Tidepool export.
}
\description{
Check if a file corresponds to a file from a Tidepool export.
}
\keyword{internal}
