% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.is_yourloops_format.aidR}
\alias{.is_yourloops_format.aidR}
\title{Check if a file comes from a YourLoops export}
\usage{
.is_yourloops_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file comes from a YourLoops export.
}
\description{
Check if a file comes from a YourLoops export
}
\keyword{internal}
