% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.match_carelink_BWZ_bolus.aidR}
\alias{.match_carelink_BWZ_bolus.aidR}
\title{Match bolus wizard entries to bolus entries from a CareLink export}
\usage{
.match_carelink_BWZ_bolus.aidR(bolus, bwz)
}
\arguments{
\item{bolus}{A data frame containing bolus entries.}

\item{bwz}{A data frame containing bolus wizard entries.}
}
\value{
A data frame containing bolus entries, with an extra column indicating the index of a matching bolus wizard entry.
}
\description{
Match bolus wizard entries to bolus entries from a CareLink export
}
\keyword{internal}
