% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.match_types_before_merging.aidR}
\alias{.match_types_before_merging.aidR}
\title{Matches the column types of multiple data frames to have matching types.}
\usage{
.match_types_before_merging.aidR(dfs_new)
}
\arguments{
\item{dfs_new}{A list of data frames.}
}
\value{
A list of data frames, with harmonized types.
}
\description{
Matches the column types of multiple data frames to have matching types.
}
\keyword{internal}
