% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_merge_per_id.R
\name{.merge_per_id.aidR}
\alias{.merge_per_id.aidR}
\title{Merge all data available for one individual, separately per data type}
\usage{
.merge_per_id.aidR(data_id)
}
\arguments{
\item{data_id}{A list with all data found for one individual, with one entry
per parsed file.}
}
\value{
A list with all data found for one individual, merged for each data
type and sorted in time. Identical entries and entries where all measurement columns are NA are removed.
}
\description{
Merge all data available for one individual, separately per data type
}
\keyword{internal}
