% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.open_carelink_file.aidR}
\alias{.open_carelink_file.aidR}
\title{Open a CareLink file}
\usage{
.open_carelink_file.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be opened.}
}
\value{
A data frame with a single column \code{V1}, holding each line of the
file as a semicolon-separated string.
}
\description{
Works if the file is in xlsx or csv format, throws an error otherwise.
}
\keyword{internal}
