% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.read_glooko_file.aidR}
\alias{.read_glooko_file.aidR}
\title{Read the data from a Glooko file}
\usage{
.read_glooko_file.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be parsed.}
}
\value{
A named list with a single entry, named after the data type of the
file and holding a data frame with formatted timestamps and numeric columns.
\code{NULL} if the file is empty. Throws an error if the data type of a
non-empty file could not be determined.
}
\description{
Read the data from a Glooko file
}
\keyword{internal}
