% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.read_tandem_source_files.aidR}
\alias{.read_tandem_source_files.aidR}
\title{Read a file from a Tandem Source export}
\usage{
.read_tandem_source_files.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename of the file to be read.}
}
\value{
A data frame containing the contents of the specified file.
}
\description{
Read a file from a Tandem Source export
}
\keyword{internal}
