% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.read_tidepool_sheet.aidR}
\alias{.read_tidepool_sheet.aidR}
\title{Read a sheet from a Tidepool Excel file}
\usage{
.read_tidepool_sheet.aidR(filename, sheet)
}
\arguments{
\item{filename}{A character string giving the path to the Excel file.}

\item{sheet}{A character string giving the name of the sheet to read.}
}
\value{
A data frame containing the contents of the specified sheet, or
\code{NULL} if the sheet does not exist in the file.
}
\description{
Read a sheet from a Tidepool Excel file
}
\keyword{internal}
