% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.read_yourloops_files.aidR}
\alias{.read_yourloops_files.aidR}
\title{Read a file from a YourLoops export}
\usage{
.read_yourloops_files.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename of the file to be read.}
}
\value{
A data frame containing the contents of the file.
}
\description{
Read a file from a YourLoops export
}
\keyword{internal}
