% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_conversions.R
\name{get_iglu_format}
\alias{get_iglu_format}
\title{Convert CGM data to a format compatible with the R-package iglu.}
\usage{
get_iglu_format(data)
}
\arguments{
\item{data}{A list with all data found for one or multiple individuals.}
}
\value{
A data frame compatible with the R-package iglu.
}
\description{
Convert CGM data to a format compatible with the R-package iglu.
}
