% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipAidData.R
\name{prepAidData}
\alias{prepAidData}
\title{Unzip and process Data}
\usage{
prepAidData(filepath = "rawdata/WITS_trade.zip", dataset = "WITS")
}
\arguments{
\item{filepath}{A character string of either a folder or zipfile (including suffix) including the data.}

\item{dataset}{The Dataset chosen for download. Options currently include "AidDataCore","China", "worldbank", .}
}
\value{
A data frame with the output data (further cleaning to be added.)
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
\dontrun{
df <- 
downloadRawAidData(dataset = "China") \%>\% 
prepAidData(filepath = .,
           dataset = "China")
head(df)

df <- 
 prepAidData(filepath = "AidDataChina.zip",
             dataset = "China")
head(df)
}
}
