% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{read_carelink}
\alias{read_carelink}
\title{Read file from a Carelink export}
\usage{
read_carelink(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Carelink export.}
}
\value{
An instance of class \code{carelink}.
}
\description{
Read file from a Carelink export
}
