% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01h_dexcom_clarity.R
\name{read_dexcom_clarity}
\alias{read_dexcom_clarity}
\title{Read a file from a Dexcom clarity export}
\usage{
read_dexcom_clarity(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string with the filename of the file to be read.}
}
\value{
An instance of class \code{dexcom_clarity}, wrapping a named list with
the CGM data.
}
\description{
Read a file from a Dexcom clarity export
}
\keyword{internal}
