% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{read_glooko}
\alias{read_glooko}
\title{Read file from a Glooko export}
\usage{
read_glooko(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Glooko export.}
}
\value{
An instance of class \code{glooko}. NULL if the data was not parsed.
}
\description{
Read file from a Glooko export
}
