% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{read_mylife}
\alias{read_mylife}
\title{Read a file from a mylife export}
\usage{
read_mylife(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string with the filename of the file to be read.}
}
\value{
An instance of class \code{mylife}, wrapping a named list with one
entry per entry type found in the file. \code{NULL} if the file is empty.
}
\description{
Read a file from a mylife export
}
\keyword{internal}
