% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{read_tandem_source}
\alias{read_tandem_source}
\title{Read a file from a Tandem Source export}
\usage{
read_tandem_source(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string with the filename of the file to be read.}
}
\value{
An instance of class \code{tandem_source}, wrapping a named list with
the data of one data type.
}
\description{
Read a file from a Tandem Source export
}
\keyword{internal}
