% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{read_tidepool}
\alias{read_tidepool}
\title{Read file from a Tidepool export}
\usage{
read_tidepool(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Tidepool export.}
}
\value{
An instance of class \code{tidepool}.
}
\description{
Read file from a Tidepool export
}
