% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{read_yourloops}
\alias{read_yourloops}
\title{Read file from a Yourloops export}
\usage{
read_yourloops(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Yourloops export.}
}
\value{
An instance of class \code{yourloops}. NULL if the data was not parsed.
}
\description{
Read file from a Yourloops export
}
