% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_demonstration.R
\name{testLuca}
\alias{testLuca}
\title{Dirichlet-sample of a multinomial election poll}
\usage{
testLuca(par1, par2)
}
\arguments{
\item{par1}{parameter1}

\item{par2}{parameter2}
}
\value{
A data frame containing \code{n} rows of blabla.
}
\description{
Description
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
x <- c(1,2,3)
mean(x)
}
