% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_missing.R
\name{this_or_default_value}
\alias{this_or_default_value}
\title{Returns given object if it exists, otherwise a provided default value}
\usage{
this_or_default_value(x, default_value)
}
\arguments{
\item{x}{An object, list, or vector}

\item{default_value}{A default value when object does not exist}
}
\value{
Object or default value
}
\description{
Returns given object if it exists, otherwise a provided default value
}
\examples{
one_string <- "hello"
two_strings <- c(one_string, one_string)

# returns the given object
this_or_default_value(one_string, "default")
this_or_default_value(two_strings, "default")

# returns the provided default value ("default" here)
this_or_default_value(NA, "default")
this_or_default_value(NULL, "default")

# returns c("default", "default") - function is vectorized
this_or_default_value(c(NA, NA), "default")
}
