% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_missing.R
\name{this_or_na}
\alias{this_or_na}
\title{Returns given object if it exists, otherwise empty string}
\usage{
this_or_na(x)
}
\arguments{
\item{x}{An object, list, or vector}
}
\value{
Object or NA
}
\description{
Returns given object if it exists, otherwise empty string
}
\examples{
one_string <- "hello"
two_strings <- c(one_string, one_string)

# returns the given object
this_or_na(one_string)
this_or_na(two_strings)

# returns NA
this_or_na(NA)
this_or_na(NULL)

# returns c(NA, NA) - function is vectorized
this_or_na(c(NA, NA))
}
