% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_munging.R
\name{tobasictext}
\alias{tobasictext}
\title{Clean up text that has extra whitespace and html elements}
\usage{
tobasictext(txt)
}
\arguments{
\item{txt}{Input text}
}
\value{
Text
}
\description{
Clean up text that has extra whitespace and html elements
}
\examples{
x <- ("<p>This is some text.\n                    This is <b>bold!</b></p>")
tobasictext(x)
}
