% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aidR-package.R
\docType{package}
\name{aidR-package}
\alias{aidR}
\alias{aidR-package}
\title{aidR: Parsing and standardizing file formats for AID data}
\description{
Parses data from automated insulin delivery (AID) systems from various platforms and allows to merge and standardize CGM, basal, bolus and carbohydrate data.
}
\author{
\strong{Maintainer}: Aritz Lizoain \email{aritz.lizoain@dcberne.com}

Authors:
\itemize{
  \item Aritz Lizoain \email{aritz.lizoain@dcberne.com}
  \item Madleina Caduff \email{madleina.caduff@dcberne.com}
}

}
\keyword{internal}
