% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_checks.R
\name{check_completeness}
\alias{check_completeness}
\title{Check if data for different types are available for at least one day}
\usage{
check_completeness(
  id,
  data_id,
  necessary_types = c("cgm", "basal", "bolus", "carbs")
)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data_id}{A list with all data found for one individual.}

\item{necessary_types}{A character vector with type names for which data should be available.}
}
\value{
No return value, called for side effects. Issues a warning for every
data type with missing data.
}
\description{
Check if data for different types are available for at least one day
}
