% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{clean.carelink}
\alias{clean.carelink}
\title{Format and clean the Carelink data to keep relevant columns only.}
\usage{
\method{clean}{carelink}(data, id, ...)
}
\arguments{
\item{data}{An instance of class \code{carelink}, wrapping a named list with data of different types.}

\item{id}{Character or numeric participant identifier.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with names \code{cgm}, \code{basal}, \code{bolus}, and \code{carbs}, and other names for data of other types.
}
\description{
Format and clean the Carelink data to keep relevant columns only.
}
