% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{clean.glooko}
\alias{clean.glooko}
\title{Format and clean the Glooko data to keep relevant columns only}
\usage{
\method{clean}{glooko}(data, id, ...)
}
\arguments{
\item{data}{An instance of class \code{glooko}, wrapping a named list with data of a particular data type.}

\item{id}{Character or numeric participant identifier.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with the cleaned data, named according to the data type
of \code{data}: \code{cgm}, \code{basal}, \code{bolus} and \code{carbs}
(both are extracted from the bolus file), \code{total_basal} and
\code{total_bolus} (both are extracted from the daily aggregated insulin
file) or \code{SMBG}. \code{NULL} for all other data types, which are not
cleaned.
}
\description{
Format and clean the Glooko data to keep relevant columns only
}
