% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_common_functions.R
\name{clean}
\alias{clean}
\title{Generic function for cleaning data}
\usage{
clean(data, ...)
}
\arguments{
\item{data}{The data to clean.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic function for cleaning data
}
