% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{clean.tandem_source}
\alias{clean.tandem_source}
\title{Format and clean the Tandem Source data to keep relevant columns only.}
\usage{
\method{clean}{tandem_source}(data, ...)
}
\arguments{
\item{data}{An instance of class \code{tandem_source}, wrapping a named list with data of a particular data type.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with names \code{cgm}, \code{basal}, \code{bolus}, or \code{carbs}, or other names if data is of another type.
}
\description{
Format and clean the Tandem Source data to keep relevant columns only.
}
