% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{clean.yourloops}
\alias{clean.yourloops}
\title{Format and clean the Yourloops data to keep relevant columns only.}
\usage{
\method{clean}{yourloops}(data, id, ...)
}
\arguments{
\item{data}{An instance of class \code{yourloops}, wrapping a named list with data of a particular data type.}

\item{id}{Character or numeric participant identifier.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list with names \code{cgm}, \code{basal}, \code{bolus}, or \code{carbs}, or other names if data is of another type.
}
\description{
Format and clean the Yourloops data to keep relevant columns only.
}
