% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.check_if_missing_columns_with_data.aidR}
\alias{.check_if_missing_columns_with_data.aidR}
\title{Throws if there are columns with non-NA entries in a data frame that are not present in subset.
Safety check to ensure no relevant columns are missed.}
\usage{
.check_if_missing_columns_with_data.aidR(df, subset)
}
\arguments{
\item{df}{The full data frame.}

\item{subset}{The subsetted data frame.}
}
\value{
Called for its side effects. Returns \code{NULL} invisibly.
}
\description{
Throws if there are columns with non-NA entries in a data frame that are not present in subset.
Safety check to ensure no relevant columns are missed.
}
\keyword{internal}
