% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.clean_carelink_basal.aidR}
\alias{.clean_carelink_basal.aidR}
\title{Format and clean basal entries of a Carelink export.}
\usage{
.clean_carelink_basal.aidR(basal)
}
\arguments{
\item{basal}{A data frame containing the basal data from a Carelink file.}
}
\value{
A data frame containing the formatted and cleaned basal data.
}
\description{
Format and clean basal entries of a Carelink export.
}
\keyword{internal}
