% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.clean_carelink_basal.aidR}
\alias{.clean_carelink_basal.aidR}
\title{Format and clean basal data from a CareLink export}
\usage{
.clean_carelink_basal.aidR(id, basal)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{basal}{A data frame containing the basal data from a CareLink file.}
}
\value{
A data frame containing the formatted and cleaned basal data.
}
\description{
Format and clean basal data from a CareLink export
}
\keyword{internal}
