% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.clean_carelink_bolus.aidR}
\alias{.clean_carelink_bolus.aidR}
\title{Format and clean bolus entries of a Carelink export.}
\usage{
.clean_carelink_bolus.aidR(id, bolus)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bolus}{A data frame containing the bolus data from a Carelink file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus entries of a Carelink export.
}
\keyword{internal}
