% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{.clean_mylife_bolus.aidR}
\alias{.clean_mylife_bolus.aidR}
\title{Format and clean bolus entries of a Mylife export.}
\usage{
.clean_mylife_bolus.aidR(bolus)
}
\arguments{
\item{bolus}{A data frame containing the bolus data from a Mylife file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus entries of a Mylife export.
}
\keyword{internal}
