% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{.clean_mylife_carbs.aidR}
\alias{.clean_mylife_carbs.aidR}
\title{Format and clean carbohydrate entries of a Mylife export.}
\usage{
.clean_mylife_carbs.aidR(id, carbs)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{carbs}{A data frame containing the carbohydrate data from a Mylife file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data.
}
\description{
Format and clean carbohydrate entries of a Mylife export.
}
\keyword{internal}
