% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.convert_BG_units_carelink.aidR}
\alias{.convert_BG_units_carelink.aidR}
\title{Converts blood glucose units to mg/dL in a Carelink file.}
\usage{
.convert_BG_units_carelink.aidR(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame, with mmol/L converted to mg/dL
}
\description{
Converts blood glucose units to mg/dL in a Carelink file.
}
\keyword{internal}
