% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.convert_excel_datetime.aidR}
\alias{.convert_excel_datetime.aidR}
\title{Convert timestamps from an Excel file formatted as fractional days}
\usage{
.convert_excel_datetime.aidR(date_val, time_val)
}
\arguments{
\item{date_val}{The fractional value for the date.}

\item{time_val}{The fractional value for the time of day.}
}
\value{
A timestamp in POSIXct, or \code{NULL} if the values are not numeric
and must be parsed as strings instead.
}
\description{
Convert timestamps from an Excel file formatted as fractional days
}
\keyword{internal}
