% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_conversions.R
\name{.diax_metadata_time.aidR}
\alias{.diax_metadata_time.aidR}
\title{Get metadata explanation for timestamps (depending on local time, with or without timestamp)}
\usage{
.diax_metadata_time.aidR(data_id)
}
\arguments{
\item{data_id}{A list with the standardized aidR data frames present for
one individual, named by aidR type (\code{cgm}, \code{bolus}, ...).}
}
\value{
A named list to be used as the \code{metadata} entry for time of the DIAX list.
}
\description{
Get metadata explanation for timestamps (depending on local time, with or without timestamp)
}
\keyword{internal}
