% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_conversions.R
\name{.diax_series_map.aidR}
\alias{.diax_series_map.aidR}
\title{Mapping from aidR data types to DIAX field names.}
\usage{
.diax_series_map.aidR()
}
\value{
A named list, mapping aidR type names to DIAX field names.
}
\description{
Mapping from aidR data types to DIAX field names.
}
\keyword{internal}
