% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.extract_estimated_absorption_duration_tidepool.aidR}
\alias{.extract_estimated_absorption_duration_tidepool.aidR}
\title{Extract estimated absorption duration from json string given in Bolus Calculator sheet}
\usage{
.extract_estimated_absorption_duration_tidepool.aidR(row, data)
}
\arguments{
\item{row}{The current row index to parse}

\item{data}{The full data frame}
}
\value{
A numeric value with the estimated absorption duration, NA if not found.
}
\description{
Extract estimated absorption duration from json string given in Bolus Calculator sheet
}
\keyword{internal}
