% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_alarms_glooko.aidR}
\alias{.format_alarms_glooko.aidR}
\title{Format and clean alarm data from Glooko.}
\usage{
.format_alarms_glooko.aidR(data)
}
\arguments{
\item{data}{A data frame containing the alarm data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned alarm data.
}
\description{
Format and clean alarm data from Glooko.
}
\keyword{internal}
