% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.format_basal_tandem_source.aidR}
\alias{.format_basal_tandem_source.aidR}
\title{Format and clean basal data from Tandem Source}
\usage{
.format_basal_tandem_source.aidR(basal)
}
\arguments{
\item{basal}{A data frame containing the basal data from a Tandem Source file.}
}
\value{
A data frame containing the formatted and cleaned basal data.
}
\description{
Format and clean basal data from Tandem Source
}
\keyword{internal}
