% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_basal_tidepool.aidR}
\alias{.format_basal_tidepool.aidR}
\title{Format and clean basal data from Tidepool}
\usage{
.format_basal_tidepool.aidR(basal)
}
\arguments{
\item{basal}{A data frame containing the basal data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned basal data.
}
\description{
Format and clean basal data from Tidepool
}
\keyword{internal}
