% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.format_bg_tandem_source.aidR}
\alias{.format_bg_tandem_source.aidR}
\title{Format and clean SMBG data from Tandem Source}
\usage{
.format_bg_tandem_source.aidR(id, bg)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bg}{A data frame containing the SMBG data from a Tandem Source file.}
}
\value{
A data frame containing the formatted and cleaned SMBG data.
}
\description{
Format and clean SMBG data from Tandem Source
}
\keyword{internal}
