% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_bolus_calculator_tidepool.aidR}
\alias{.format_bolus_calculator_tidepool.aidR}
\title{Format and clean carbohydrates from bolus calculator data from Tidepool}
\usage{
.format_bolus_calculator_tidepool.aidR(bolus_calculator)
}
\arguments{
\item{bolus_calculator}{A data frame containing the bolus calculator data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrates from bolus calculator data.
}
\description{
Format and clean carbohydrates from bolus calculator data from Tidepool
}
\keyword{internal}
