% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_bolus_glooko.aidR}
\alias{.format_bolus_glooko.aidR}
\title{Format and clean bolus data from Glooko.}
\usage{
.format_bolus_glooko.aidR(data)
}
\arguments{
\item{data}{A data frame containing the bolus data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus data from Glooko.
}
\keyword{internal}
