% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.format_bolus_tandem_source.aidR}
\alias{.format_bolus_tandem_source.aidR}
\title{Format and clean bolus data from Tandem Source}
\usage{
.format_bolus_tandem_source.aidR(id, bolus)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{bolus}{A data frame containing the bolus data from a Tandem Source file.}
}
\value{
A data frame containing the formatted and cleaned bolus data.
}
\description{
Format and clean bolus data from Tandem Source
}
\keyword{internal}
