% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_cgm_yourloops.aidR}
\alias{.format_cgm_yourloops.aidR}
\title{Format and clean CGM data from Yourloops}
\usage{
.format_cgm_yourloops.aidR(cgm)
}
\arguments{
\item{cgm}{A data frame containing the CGM data from a Yourloops file.}
}
\value{
A data frame containing the formatted and cleaned CGM data.
}
\description{
Format and clean CGM data from Yourloops
}
\keyword{internal}
