% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_food_glooko.aidR}
\alias{.format_food_glooko.aidR}
\title{Format and clean food data from Glooko.}
\usage{
.format_food_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the food data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned food data.
}
\description{
Format and clean food data from Glooko.
}
\keyword{internal}
