% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_food_tidepool.aidR}
\alias{.format_food_tidepool.aidR}
\title{Format and clean carbohydrate data from the Food sheet of a Tidepool export}
\usage{
.format_food_tidepool.aidR(id, food)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{food}{A data frame containing the food data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned carbohydrate data.
}
\description{
Format and clean carbohydrate data from the Food sheet of a Tidepool export
}
\keyword{internal}
