% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_medication_glooko.aidR}
\alias{.format_medication_glooko.aidR}
\title{Format and clean medication data from Glooko.}
\usage{
.format_medication_glooko.aidR(data)
}
\arguments{
\item{data}{A data frame containing the medication data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned medication data.
}
\description{
Format and clean medication data from Glooko.
}
\keyword{internal}
