% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_notes_glooko.aidR}
\alias{.format_notes_glooko.aidR}
\title{Format and clean notes data from Glooko.}
\usage{
.format_notes_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the notes data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned notes data.
}
\description{
Format and clean notes data from Glooko.
}
\keyword{internal}
