% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_rescuecarbs_yourloops.aidR}
\alias{.format_rescuecarbs_yourloops.aidR}
\title{Format and clean rescuecarbs data from Yourloops}
\usage{
.format_rescuecarbs_yourloops.aidR(id, carbs)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{carbs}{A data frame containing the rescuecarbs data from a Yourloops file.}
}
\value{
A data frame containing the formatted and cleaned rescuecarbs data.
}
\description{
Format and clean rescuecarbs data from Yourloops
}
\keyword{internal}
