% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_SMBG_glooko.aidR}
\alias{.format_SMBG_glooko.aidR}
\title{Format and clean SMBG data from a Glooko export}
\usage{
.format_SMBG_glooko.aidR(id, data)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{data}{A data frame containing the SMBG data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned SMBG data.
}
\description{
Format and clean SMBG data from a Glooko export
}
\keyword{internal}
