% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_tidepool.R
\name{.format_SMBG_tidepool.aidR}
\alias{.format_SMBG_tidepool.aidR}
\title{Format and clean SMBG data from Tidepool}
\usage{
.format_SMBG_tidepool.aidR(id, SMBG)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{cgm}{A data frame containing the SMBG data from a Tidepool file.}
}
\value{
A data frame containing the formatted and cleaned SMBG data.
}
\description{
Format and clean SMBG data from Tidepool
}
\keyword{internal}
