% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.format_timestamp_yourloops.aidR}
\alias{.format_timestamp_yourloops.aidR}
\title{Format timestamp to get local (wall-clock) time of a Yourloops export}
\usage{
.format_timestamp_yourloops.aidR(file)
}
\arguments{
\item{file}{A data frame.}
}
\value{
A data frame where the timezone has been formatted to local time.
}
\description{
Format timestamp to get local (wall-clock) time of a Yourloops export
}
\keyword{internal}
