% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.format_total_insulin_glooko.aidR}
\alias{.format_total_insulin_glooko.aidR}
\title{Format and clean total insulin data from Glooko.}
\usage{
.format_total_insulin_glooko.aidR(data)
}
\arguments{
\item{data}{A data frame containing the total insulin data from a Glooko file.}
}
\value{
A data frame containing the formatted and cleaned total insulin data.
}
\description{
Format and clean total insulin data from Glooko.
}
\keyword{internal}
