% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_alerts.aidR}
\alias{.get_carelink_alerts.aidR}
\title{Parse all entries corresponding to alerts in a CareLink export}
\usage{
.get_carelink_alerts.aidR(df)
}
\arguments{
\item{df}{A data frame obtained from reading a CareLink file.}
}
\value{
A data frame with the corresponding entries.
}
\description{
Parse all entries corresponding to alerts in a CareLink export
}
\keyword{internal}
