% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_CGM.aidR}
\alias{.get_carelink_CGM.aidR}
\title{Parse all entries corresponding to CGM data in a Carelink export.}
\usage{
.get_carelink_CGM.aidR(df)
}
\arguments{
\item{df}{A data frame obtained from reading a Carelink file.}
}
\value{
A data frame containing the target entries
}
\description{
Parse all entries corresponding to CGM data in a Carelink export.
}
\keyword{internal}
