% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.get_carelink_CGM.aidR}
\alias{.get_carelink_CGM.aidR}
\title{Parse all entries corresponding to CGM data in a CareLink export}
\usage{
.get_carelink_CGM.aidR(df, sensor_name)
}
\arguments{
\item{df}{A data frame obtained from reading a CareLink file.}

\item{sensor_name}{A character string with the name of the sensor, \code{NA} if unknown.}
}
\value{
A data frame with the corresponding entries.
}
\description{
Parse all entries corresponding to CGM data in a CareLink export
}
\keyword{internal}
