% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.get_tandem_source_file_lookup.aidR}
\alias{.get_tandem_source_file_lookup.aidR}
\title{A lookup with Tandem Source file characteristics}
\usage{
.get_tandem_source_file_lookup.aidR()
}
\value{
A list with data types and their expected headers.
}
\description{
A lookup with Tandem Source file characteristics
}
\keyword{internal}
