% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{.get_tandem_source_file_type.aidR}
\alias{.get_tandem_source_file_type.aidR}
\title{Get file type of a Tandem Source export.}
\usage{
.get_tandem_source_file_type.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A character string if file type could be determined, NULL there was no match. If multiple files types matched, an error is thrown.
}
\description{
Get file type of a Tandem Source export.
}
\keyword{internal}
