% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.get_yourloops_file_type.aidR}
\alias{.get_yourloops_file_type.aidR}
\title{Get file type of a Yourloops export.}
\usage{
.get_yourloops_file_type.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A character string if file type could be determined, NULL there was no match. If multiple files types matched, an error is thrown.
}
\description{
Get file type of a Yourloops export.
}
\keyword{internal}
