% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.is_carelink_format.aidR}
\alias{.is_carelink_format.aidR}
\title{Check if a file corresponds to a file from a Carelink export.}
\usage{
.is_carelink_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file is a file from a Carelink export.
}
\description{
Check if a file corresponds to a file from a Carelink export.
}
\keyword{internal}
