% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.is_glooko_file_type.aidR}
\alias{.is_glooko_file_type.aidR}
\title{Check if a Glooko file corresponds to a particular data type.}
\usage{
.is_glooko_file_type.aidR(
  filename,
  expected_num_cols,
  numeric_cols,
  non_numeric_cols = c(),
  expected_keyword = NULL
)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}

\item{expected_num_cols}{The expected number of columns of that file.}

\item{numeric_cols}{A numeric vector corresponding to the column indices that contain numeric values.}

\item{non_numeric_cols}{A numeric vector corresponding to the column indices that should not contain numeric values.}

\item{expected_keyword}{A list with keys \code{name} and \code{col_ix}, corresponding to a keyword that should be present in a particular column.}
}
\value{
A logical value: \code{TRUE} if the file matches the data type format.
}
\description{
Check if a Glooko file corresponds to a particular data type.
}
\keyword{internal}
