% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{.is_mylife_format.aidR}
\alias{.is_mylife_format.aidR}
\title{Check if a file corresponds to a file from a Mylife export.}
\usage{
.is_mylife_format.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be checked.}
}
\value{
A logical value: \code{TRUE} if the file is a file from a Mylife export.
}
\description{
Check if a file corresponds to a file from a Mylife export.
}
\keyword{internal}
