% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.is_numeric_col.aidR}
\alias{.is_numeric_col.aidR}
\title{Check if a vector can be converted to numeric values}
\usage{
.is_numeric_col.aidR(vec)
}
\arguments{
\item{vec}{Character vector to be checked.}
}
\value{
A logical value: \code{TRUE} if the vector can be converted to numeric values.
}
\description{
Check if a vector can be converted to numeric values
}
\keyword{internal}
