% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.open_carelink_file.aidR}
\alias{.open_carelink_file.aidR}
\title{Open a Carelink file. Works if file is in xlsx or csv format.
Throws error if file extension does not match either.}
\usage{
.open_carelink_file.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be opened.}
}
\value{
A data frame.
}
\description{
Open a Carelink file. Works if file is in xlsx or csv format.
Throws error if file extension does not match either.
}
\keyword{internal}
