% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_carelink.R
\name{.read_carelink_raw.aidR}
\alias{.read_carelink_raw.aidR}
\title{Read file from a Carelink export}
\usage{
.read_carelink_raw.aidR(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Glooko export.}
}
\value{
A list of three data frames containing (1) many different information, (2) daily aggregated insulin and (3) CGM measurement data.
}
\description{
Read file from a Carelink export
}
\keyword{internal}
