% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_glooko.R
\name{.read_glooko_file.aidR}
\alias{.read_glooko_file.aidR}
\title{Read the data from a Glooko file.}
\usage{
.read_glooko_file.aidR(filename)
}
\arguments{
\item{filename}{Character string with the filename to be parsed.}
}
\value{
A data frame with the content found in the file, with formatted timestamps and numeric columns. NULL if the file could not be parsed
}
\description{
Read the data from a Glooko file.
}
\keyword{internal}
