% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{.read_yourloops_files.aidR}
\alias{.read_yourloops_files.aidR}
\title{Read a file from a Yourloops export}
\usage{
.read_yourloops_files.aidR(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file.}
}
\value{
A data frame containing the contents of the specified sheet, or
\code{NULL} if the sheet does not exist in the file.
}
\description{
Read a file from a Yourloops export
}
\keyword{internal}
