% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_filtering.R
\name{filter_CGM}
\alias{filter_CGM}
\title{Filter CGM values}
\usage{
filter_CGM(
  data_id,
  truncate_min_cgm = 40,
  truncate_max_cgm = 400,
  min_minutes_difference_cgm = 5
)
}
\arguments{
\item{data_id}{A list with all data found for one individual. Filtering is
only performed if the CGM data was cleaned and standardized.}

\item{truncate_min_cgm}{CGM values smaller than this threshold will be truncated. Default 40 mg/dL, if NULL, no filtering is performed.}

\item{truncate_max_cgm}{CGM values larger than this threshold will be truncated. Default 400 mg/dL, if NULL, no filtering is performed.}

\item{min_minutes_difference_cgm}{CGM values are downsampled to this time interval (in minutes). Default 5 minutes, if NULL, no downsampling is performed.}
}
\value{
A list with all data found for one individual, filtered, if necessary.
}
\description{
Filter CGM values
}
