% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_merge_all.R
\name{merge_all}
\alias{merge_all}
\title{Merge data available for all individuals, separately per data type.}
\usage{
merge_all(data)
}
\arguments{
\item{data}{A list with data found for all individual.}
}
\value{
A list with data found for all individual, merged for each data type.
}
\description{
Merge data available for all individuals, separately per data type.
}
