% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_mylife.R
\name{read_mylife}
\alias{read_mylife}
\title{Read file from a Mylife export}
\usage{
read_mylife(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Mylife export.}
}
\value{
An instance of class \code{mylife}.
}
\description{
Read file from a Mylife export
}
