% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_tandem_source.R
\name{read_tandem_source}
\alias{read_tandem_source}
\title{Read file from a Tandem Source export}
\usage{
read_tandem_source(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string corresponding to the filename of the Tandem Source export.}
}
\value{
An instance of class \code{tandem_source}. NULL if the data was not parsed.
}
\description{
Read file from a Tandem Source export
}
