% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_yourloops.R
\name{read_yourloops}
\alias{read_yourloops}
\title{Read a file from a YourLoops export}
\usage{
read_yourloops(id, filename)
}
\arguments{
\item{id}{Character or numeric participant identifier.}

\item{filename}{Character string with the filename of the file to be read.}
}
\value{
An instance of class \code{yourloops}, wrapping a named list with the
data of one data type.
}
\description{
Read a file from a YourLoops export
}
\keyword{internal}
