% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_missing.R
\name{this_exists}
\alias{this_exists}
\title{Test if an object exists}
\usage{
this_exists(x)
}
\arguments{
\item{x}{Any object}
}
\value{
TRUE or FALSE
}
\description{
Test if an object exists
}
\examples{
# these are TRUE
this_exists(4)
this_exists(TRUE)
this_exists(FALSE)
this_exists("a word")
this_exists(c("a", "b"))
this_exists(0)

# these are FALSE
this_exists(c())
this_exists(list())
this_exists(NULL)
this_exists(NA)
}
