% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_missing.R
\name{this_is_singular}
\alias{this_is_singular}
\title{Tests if all entries in a list or vector are the same}
\usage{
this_is_singular(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
TRUE or FALSE
}
\description{
Only returns TRUE if there are no NA's or NULL's in the list or vector
}
\examples{
# these are TRUE
this_is_singular("a")
this_is_singular(c("a", "a"))
this_is_singular(list("a", "a"))

# these are FALSE
this_is_singular(c("a", "b"))
this_is_singular(c("a", NA))
this_is_singular(list("a", NULL))
}
