% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_missing.R
\name{this_or_empty_string}
\alias{this_or_empty_string}
\title{Returns given object if it exists, otherwise empty string}
\usage{
this_or_empty_string(x)
}
\arguments{
\item{x}{An object, list, or vector}
}
\value{
Object or empty string ("")
}
\description{
Returns given object if it exists, otherwise empty string
}
\examples{
one_string <- "hello"
two_strings <- c(one_string, one_string)

# returns the given object
this_or_empty_string(one_string)
this_or_empty_string(two_strings)

# returns an empty string ("")
this_or_empty_string(NA)
this_or_empty_string(NULL)

# returns c("", "") - function is vectorized
this_or_empty_string(c(NA, NA))
}
